express-xml-domparser XLM DOM parse middleware
Some applications still either need or want to process XML
sent to it as XML and not as javascript objects.  This is
connect/express middleware to allow http requests that contain
raw XML to be parsed into a DOM.  It uses the libxmljs npm
package as its parser, and exposes 2 new properties on the
req http.ClientRequest object.  They are:
  -req.rawBody: the raw XML data read in before any processing
  -req.xmlDom: the libxmljs DOM of the document